pyproject.toml
[project]
name            = "zrm"
version         = "3.0.0"
description     = "A minimal, single-file communication middleware built on Zenoh"
readme          = "README.md"
requires-python = ">=3.11"

authors = [{ name = "JafarAbdi", email = "jafar.uruc@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: System :: Networking",
]

dependencies = ["eclipse-zenoh>=1.6.2", "protobuf>=6.33.0"]

[project.urls]
Homepage   = "https://github.com/JafarAbdi/zrm"
Repository = "https://github.com/JafarAbdi/zrm"
Issues     = "https://github.com/JafarAbdi/zrm/issues"
CI         = "https://github.com/JafarAbdi/zrm/actions"

[dependency-groups]
dev = ["pre-commit>=4.2.0", "pytest>=8.4.1"]

[project.scripts]
zrm-session = "zrm.cli.session:main"
zrm-topic   = "zrm.cli.topic:main"
zrm-service = "zrm.cli.service:main"
zrm-proto   = "zrm.cli.proto:main"

[build-system]
requires      = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
